only look at ApiCompat on Windows, add retry for previous version download

+++ build/PreBuild.ps1
@@ -1,8 +1,16 @@
 param([string]$package, [string]$version)
+
 $tempFile = "..\LastMajorVersionBinaries\$package.$version.txt"
-if ((Test-Path -Path $tempFile) -OR (Test-Path -Path "..\LastMajorVersionBinaries\$package.$version.zip"))
+$retryCount = 0
+while ((Test-Path -Path $tempFile) -and ($retryCount -Lt 10))
 {
-    Write-Debug "Previous package version already exists or retrieval is in progress"
+    Write-Host "Previous version retrieval in progress, waiting..."
+    $retryCount++
+    Start-Sleep -s 1
+}
+if (Test-Path -Path "..\LastMajorVersionBinaries\$package.$version.zip")
+{
+    Write-Debug "Previous package version already exists"
 }
 else
 {
@@ -11,5 +19,9 @@ else
 
     Invoke-WebRequest -Uri https://www.nuget.org/api/v2/package/$package/$version -Outfile ..\LastMajorVersionBinaries\$package.$version.zip
     Expand-Archive -LiteralPath ..\LastMajorVersionBinaries\$package.$version.zip -DestinationPath ..\LastMajorVersionBinaries\$package\$version
+}
+
+if (Test-Path -Path $tempFile)
+{
     Remove-Item $tempFile
 }
